requirements.txt
Flask
gunicorn
tensorflow-cpu==2.21.0
keras==3.13.2
numpy
beautifulsoup4